Move a lot of render functionality down to presentational
