+++ frontend/src/components/CreateSeason/Presentational.tsx
@@ -1,41 +1,42 @@
+import AlertStripe from 'nav-frontend-alertstriper';
 import KnappBase from 'nav-frontend-knapper';
 import { SkjemaGruppe } from 'nav-frontend-skjema';
-import { SkjemaelementFeil } from 'nav-frontend-skjema/lib/skjemaelement-feilmelding';
 import { Sidetittel } from 'nav-frontend-typografi';
 import React from 'react';
 import { IApplicationSeason } from '../ApplicationSeason/reducer';
 import { TitleAndSpinner } from '../LoadingPageSpinner/TitleAndSpinner';
 import {
   _CREATE_NEW_SEASON,
+  _ERROR_MESSAGE,
   _EXISTING_CURRENT_SEASON,
   _NEW_APPLICATION_SEASON,
+  _ROOM_END_TOO_EARLY,
+  _SEASON_END_TOO_EARLY,
   _UPDATE_APPLICATION_SEASON,
   _UPDATE_CURRENT_SEASON,
 } from './strings';
 
 interface IProps {
-  alertPeriodEndBeforeStart?: SkjemaelementFeil;
-  alertApplicationEndBeforeStart?: SkjemaelementFeil;
   createFields: (index: number, end: number) => JSX.Element[];
-  buttonDisable: boolean;
-  alert?: JSX.Element;
+  showAlert: boolean;
   season: IApplicationSeason;
   submitSeason: () => void;
   fetched: boolean;
 }
 
+const errorObjectSeasonEndTooEarly = { feilmelding: _SEASON_END_TOO_EARLY };
+const errorObjectRoomEndTooEarly = { feilmelding: _ROOM_END_TOO_EARLY };
+
 const Presentational: React.FunctionComponent<IProps> = (props) => {
   const {
-    alertApplicationEndBeforeStart,
-    alertPeriodEndBeforeStart,
     createFields,
-    buttonDisable,
     submitSeason,
-    alert,
+    showAlert,
     season,
     fetched,
   } = props;
   const title = fetched && season.id ? _UPDATE_APPLICATION_SEASON : _NEW_APPLICATION_SEASON;
+  const { applicationPeriodEnd, applicationPeriodStart, end, start } = season;
   if (!fetched) return <TitleAndSpinner title={title} />;
 
   const createButton = (text: string) => (
@@ -48,6 +49,23 @@ const Presentational: React.FunctionComponent<IProps> = (props) => {
       {text}
     </KnappBase>
   );
+  const errorPeriodEndBeforeStart =
+    applicationPeriodEnd <= applicationPeriodStart
+      ? errorObjectSeasonEndTooEarly
+      : undefined;
+
+  const errorApplicationEndBeforeStart =
+    end <= start
+      ? errorObjectRoomEndTooEarly
+      : undefined;
+
+  const buttonDisable =
+    errorPeriodEndBeforeStart !== undefined
+    || errorApplicationEndBeforeStart !== undefined;
+
+  const alertFail = showAlert
+    ? <AlertStripe type={'advarsel'} solid={true}> {_ERROR_MESSAGE} </AlertStripe>
+    : undefined;
 
   const button = season.id > 0
     ? createButton(_UPDATE_CURRENT_SEASON)
@@ -56,14 +74,14 @@ const Presentational: React.FunctionComponent<IProps> = (props) => {
   return (
     <div id="new-application-season" className="main-content">
       <Sidetittel>{title}</Sidetittel>
-      <div id="admin-season-alert">{alert}</div>
+      <div id="admin-season-alert">{alertFail}</div>
       <div id="appSeason">
-        <SkjemaGruppe className="app-class" feil={alertPeriodEndBeforeStart}>
+        <SkjemaGruppe className="app-class" feil={errorPeriodEndBeforeStart}>
           {createFields(0, 2)}
         </SkjemaGruppe>
       </div>
       <div id="roomSeason">
-        <SkjemaGruppe className="room-class" feil={alertApplicationEndBeforeStart}>
+        <SkjemaGruppe className="room-class" feil={errorApplicationEndBeforeStart}>
           {createFields(2, 4)}
         </SkjemaGruppe>
       </div>
